Refactor form into own component to make logic simpler
