+++ app/web/features/profile/locales/de.json
@@ -60,7 +60,7 @@
     "general": "Allgemein",
     "has_housemates": "Hat Mitbewohner",
     "host_drinking": "Trinkt zu Hause",
-    "host_kids": "Hat Kinder",
+    "has_kids": "Hat Kinder",
     "house_rules": "Hausregeln",
     "housemates": "Mitbewohner",
     "kid_details": "Details zu Kindern",
