+++ app/web/features/profile/locales/en.json
@@ -54,8 +54,8 @@
     "has_housemates": "Has housemates",
     "home_facilities": "Home Facilities",
     "host_drinking": "Drinks at home",
-    "host_kids": "Has children",
-    "host_pets": "Has pets",
+    "has_kids": "Has kids",
+    "has_pets": "Has pets",
     "host_smoking": "Smokes at home",
     "hosting_preferences": "Hosting Preferences",
     "house_rules": "House rules",
