+++ app/web/features/profile/locales/fr.json
@@ -106,9 +106,9 @@
     "last_minute": "Accepter les demandes d'hébergement de dernière minute",
     "local_area": "Informations locales",
     "about_home": "À propos de mon logement",
-    "host_kids": "a des enfants",
+    "has_kids": "a des enfants",
     "host_drinking": "boit l'alcool chez soi",
-    "host_pets": "a des animaux de compagnie",
+    "has_pets": "a des animaux de compagnie",
     "general": "Section générale",
     "has_housemates": "a des colocataires",
     "hosting_preferences": "Préférences d’hébergement",
