+++ app/web/features/profile/locales/nl.json
@@ -57,8 +57,8 @@
     "general": "Algemeen",
     "has_housemates": "Heeft huisgenoten",
     "host_drinking": "Drinkt thuis",
-    "host_kids": "Heeft kinderen",
-    "host_pets": "Heeft huisdieren",
+    "has_kids": "Heeft kinderen",
+    "has_pets": "Heeft huisdieren",
     "max_guests": "Maximaal aantal gasten",
     "my_home": "Mijn huis",
     "parking": "Parkeerruimte beschikbaar",
