+++ app/web/features/profile/locales/zh_Hans.json
@@ -40,8 +40,8 @@
     "general": "常规",
     "has_housemates": "有室友",
     "host_drinking": "家中是否饮酒",
-    "host_kids": "有孩子",
-    "host_pets": "有宠物",
+    "has_kids": "有孩子",
+    "has_pets": "有宠物",
     "host_smoking": "家中是否吸烟",
     "hosting_preferences": "接待偏好",
     "house_rules": "房屋规则",
