+++ app/web/test/fixtures/users.json
@@ -51,6 +51,8 @@
     "parking": { "value": true },
     "parkingDetails": 3,
     "smokingAllowed": 1,
+    "acceptsKids": { "value": true },
+    "acceptsPets": { "value": true },
     "avatarUrl": "https://loremflickr.com/200/200?user1",
     "avatarThumbnailUrl": "https://loremflickr.com/50/50?user1",
     "badgesList": ["founder", "donor", "strong_verification"],
